Unify marketing locale routes
