--- src/app/(pages)/layout.tsx
@@ -1,21 +1,0 @@
-import { Header } from "@/components/homepage/header";
-import { Footer } from "@/components/homepage/footer";
-import { Suspense } from "react";
-
-export const dynamic = "force-static";
-
-export default function PagesLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="flex min-h-screen flex-col">
-      <Suspense fallback={<div className="h-16 w-full" />}>
-        <Header />
-      </Suspense>
-      <main className="flex-1">{children}</main>
-      <Footer />
-    </div>
-  );
-}
