--- src/app/(pages)/payment-status/page.tsx
@@ -1,50 +1,0 @@
-import { Suspense } from "react";
-import { PaymentStatusContent } from "./_components/payment-status-content";
-import { Card, CardContent } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
-import { Badge } from "@/components/ui/badge";
-import { SectionContainer } from "@/components/layout/page-container";
-import { Clock } from "lucide-react";
-
-function PaymentStatusSkeleton() {
-  return (
-    <section className="bg-background relative flex min-h-screen items-center justify-center overflow-hidden">
-      {/* Background Pattern */}
-      <div className="absolute inset-0 -z-10">
-        <div className="absolute inset-0 bg-[radial-gradient(circle_at_50%_50%,rgba(120,119,198,0.03),transparent_50%)]" />
-        <div className="absolute inset-0 bg-[linear-gradient(to_right,hsl(var(--border))_1px,transparent_1px),linear-gradient(to_bottom,hsl(var(--border))_1px,transparent_1px)] bg-[size:4rem_4rem] opacity-20" />
-      </div>
-
-      <SectionContainer className="relative">
-        {/* Status Badge Skeleton */}
-        <div className="mb-8 text-center">
-          <Badge
-            variant="outline"
-            className="border-border bg-background/50 text-muted-foreground inline-flex items-center border px-3 py-1 font-mono text-sm backdrop-blur-sm"
-          >
-            <Clock className="mr-2 h-3 w-3" />
-            LOADING_STATUS
-          </Badge>
-        </div>
-
-        <Card className="w-full text-center">
-          <CardContent className="pt-6">
-            <Skeleton className="mx-auto mb-6 h-16 w-16 rounded-full" />
-            <Skeleton className="mx-auto mb-4 h-8 w-64" />
-            <Skeleton className="mx-auto mb-8 h-4 w-80" />
-            <Skeleton className="mx-auto mb-3 h-10 w-48" />
-            <Skeleton className="mx-auto h-10 w-40" />
-          </CardContent>
-        </Card>
-      </SectionContainer>
-    </section>
-  );
-}
-
-export default function PaymentStatusPage() {
-  return (
-    <Suspense fallback={<PaymentStatusSkeleton />}>
-      <PaymentStatusContent />
-    </Suspense>
-  );
-}
