+++ src/app/[locale]/payment-status/_components/payment-status-content.tsx
