+++ src/lib/i18n/static-marketing-locale.test.ts
@@ -10,16 +10,19 @@ import {
 } from "./static-marketing-locale";
 
 describe("static marketing locale helpers", () => {
-  it("returns only target locales for localized marketing paths", () => {
-    expect(getStaticMarketingLocaleParams()).toEqual([{ locale: "zh-Hans" }]);
+  it("returns all supported locales for the unified marketing route tree", () => {
+    expect(getStaticMarketingLocaleParams()).toEqual([
+      { locale: "en" },
+      { locale: "zh-Hans" },
+    ]);
   });
 
-  it("accepts canonical target locale segments", () => {
+  it("accepts canonical supported locale segments", () => {
+    expect(resolveStaticMarketingLocale("en")).toBe("en");
     expect(resolveStaticMarketingLocale("zh-Hans")).toBe("zh-Hans");
   });
 
-  it("rejects source, alias, and unsupported locale segments", () => {
-    expect(() => resolveStaticMarketingLocale("en")).toThrow("NEXT_NOT_FOUND");
+  it("rejects alias and unsupported locale segments", () => {
     expect(() => resolveStaticMarketingLocale("zh")).toThrow("NEXT_NOT_FOUND");
     expect(() => resolveStaticMarketingLocale("fr")).toThrow("NEXT_NOT_FOUND");
   });
