+++ src/lib/i18n/static-marketing-locale.ts
@@ -1,28 +1,24 @@
 import { notFound } from "next/navigation";
 
-import {
-  SOURCE_LOCALE,
-  TARGET_LOCALES,
-  type SupportedLocale,
-} from "@/lib/config/i18n";
+import { SUPPORTED_LOCALES, type SupportedLocale } from "@/lib/config/i18n";
 import { normalizeLocaleCandidate } from "@/lib/config/i18n-routing";
 
 export type StaticMarketingLocaleParams = {
   locale: SupportedLocale;
 };
 
+export type LocalizedMarketingPageProps = {
+  params: Promise<{ locale: string }>;
+};
+
 export function getStaticMarketingLocaleParams(): StaticMarketingLocaleParams[] {
-  return TARGET_LOCALES.map((locale) => ({ locale }));
+  return SUPPORTED_LOCALES.map((locale) => ({ locale }));
 }
 
 export function resolveStaticMarketingLocale(locale: string): SupportedLocale {
   const normalizedLocale = normalizeLocaleCandidate(locale);
 
-  if (
-    !normalizedLocale ||
-    normalizedLocale === SOURCE_LOCALE ||
-    normalizedLocale !== locale
-  ) {
+  if (!normalizedLocale || normalizedLocale !== locale) {
     notFound();
   }
